cpu: deduplicate arm64 cpuInfo into a shared file
Address review feedback: cpuInfo() was identical in cpu_arm64.go and
cpu_arm64_darwin.go (the two files differ only in their init() feature
detection). Move it to cpu_arm64_info.go behind //go:build arm64 so both
platform builds share one implementation.

Verified on a Raspberry Pi 5 (linux/arm64): TestCpuInfoAllBranches and TestInfo
pass; darwin/arm64 cross-build is green.
