+++ cpu/cpu_arm64.go
@@ -19,25 +19,4 @@ func init() {
 	applyDisable(&ARM64, os.Getenv("SIMD_DISABLE"))
 }
 
-func cpuInfo() string {
-	// Report the tier the library actually runs (NEON), not SVE: there are no SVE
-	// kernels yet, so an SVE-capable host still executes the NEON path. SVE/SVE2 is
-	// annotated as detected-but-unused so the capability is still visible.
-	var base string
-	switch {
-	case ARM64.NEON && ARM64.FP16:
-		base = "ARM64 NEON+FP16"
-	case ARM64.NEON:
-		base = "ARM64 NEON"
-	default:
-		return "ARM64 (no SIMD)"
-	}
-	switch {
-	case ARM64.SVE2:
-		return base + " (SVE2 detected, unused)"
-	case ARM64.SVE:
-		return base + " (SVE detected, unused)"
-	default:
-		return base
-	}
-}
+// cpuInfo is shared with the darwin build; see cpu_arm64_info.go.
